requirements.txt
# Переменные окружения
python-dotenv==1.0.1

# Telegram-бот
pyTelegramBotAPI>=4.0.0

# LLM через OpenRouter
langchain-openai>=0.3.0

# Загрузка и разбор документов базы знаний (RAG)
python-docx>=1.1.0
pypdf>=5.1.0
cryptography>=3.1

# Векторный поиск (FAISS) и эмбеддинги через OpenRouter
faiss-cpu>=1.7.4
langchain-community>=0.3.0

# Веб-поиск
tavily-python>=0.3.0

# Агент с инструментами и циклом рассуждений
langgraph>=0.2.0